Version 2

+++ api/src/models/cotizacioncond.js
@@ -17,19 +17,19 @@ module.exports = (sequelize) => {
             allowNull: false,
         },        
         descuento: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL,
             allowNull: false,
         },        
         enganche: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL,
             allowNull: false,
         },
         meses: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL,
             allowNull: false,
         },
         interes: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL,
             allowNull: false,
         }
     },
